+++ pkg/river/vm/vm_stdlib_test.go
@@ -135,6 +135,48 @@ func TestStdlib_Nonsensitive(t *testing.T) {
 		})
 	}
 }
+func TestStdlib_StringFunc(t *testing.T) {
+	scope := &vm.Scope{
+		Variables: map[string]any{},
+	}
+
+	tt := []struct {
+		name   string
+		input  string
+		expect interface{}
+	}{
+		{"lower", `lower("String")`, "string"},
+		{"upper", `upper("string")`, "STRING"},
+		{"trimspace", `trimspace("   string \n\n")`, "string"},
+		{"trimspace+upper+trim", `lower(upper(trimspace("   String   ")))`, "string"},
+		{"split", `split("/aaa/bbb/ccc/ddd", "/")`, []string{"", "aaa", "bbb", "ccc", "ddd"}},
+		{"split+index", `split("/aaa/bbb/ccc/ddd", "/")[0]`, ""},
+		{"join+split", `join(split("/aaa/bbb/ccc/ddd", "/"), "/")`, "/aaa/bbb/ccc/ddd"},
+		{"join", `join(["foo", "bar", "baz"], ", ")`, "foo, bar, baz"},
+		{"format", `format("Hello %s", "World")`, "Hello World"},
+		{"format+int", `format("%#v", 1)`, "1"},
+		{"format+bool", `format("%#v", true)`, "true"},
+		{"format+quote", `format("%q", "hello")`, `"hello"`},
+		{"replace", `replace("Hello World", " World", "!")`, "Hello!"},
+		{"trim", `trim("?!hello?!", "!?")`, "hello"},
+		{"trim2", `trim("   hello! world.!  ", "! ")`, "hello! world."},
+		{"trimprefix", `trimprefix("helloworld", "hello")`, "world"},
+		{"trimsuffix", `trimsuffix("helloworld", "world")`, "hello"},
+	}
+
+	for _, tc := range tt {
+		t.Run(tc.name, func(t *testing.T) {
+			expr, err := parser.ParseExpression(tc.input)
+			require.NoError(t, err)
+
+			eval := vm.New(expr)
+
+			rv := reflect.New(reflect.TypeOf(tc.expect))
+			require.NoError(t, eval.Evaluate(scope, rv.Interface()))
+			require.Equal(t, tc.expect, rv.Elem().Interface())
+		})
+	}
+}
 
 func BenchmarkConcat(b *testing.B) {
 	// There's a bit of setup work to do here: we want to create a scope holding