+++ CHANGELOG.md
@@ -52,6 +52,8 @@ Main (unreleased)
 
 - Added json_path function to river stdlib. (@jkroepke)
 
+- Add `format`, `join`, `lower`, `replace`, `split`, `trimprefix`, `trimsuffix`, `trimsuffix`, `trimspace`, `upper` functions to river stdlib. (@jkroepke)
+
 - Flow UI: Add a view for listing the Agent's peers status when clustering is enabled. (@tpaschalis)
 
 - Add a new CLI command `grafana-agent convert` for converting a river file from supported formats to river. (@erikbaranowski)